requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
click==8.1.3
    # via -r requirements.in
colorama==0.4.6
    # via click
numpy==1.24.3
    # via -r requirements.in
pillow==9.5.0
    # via -r requirements.in
